backend/cart/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
mysqlclient==2.2.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
python-decouple
drf_yasg
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt
pika==1.3.0
requests